+++ CHANGELOG.md
@@ -9,6 +9,7 @@ principles.
 - Fixed: issue where deleting model redirected to the index inside the sidebar.
 - Fixed: sorting logic can now handle uuids, rather than only integers.
 - Fixed: double clicks sometimes opened up sidebar with duplicate content.
+- Fixed: adding file on create page didn't upload.
 - Changed: Field::options callback now has the model as second parameter.
 - Added: option to choose where to go to after creating model. Via `Resource::getRedirectAfterCreate()`. This can be set to null as well aka when used in sidebar.
 - Added: resource::getInstanceAttributes method to set default attributes on a model when record is not created yet